merge fix from main

+++ ccdb.py
@@ -2,7 +2,7 @@ import numpy as np
 
 def read_pic(name):
     fid = open(name, 'rb')
-    nx,ny,nz = map(np.int, np.fromfile(fid, np.uint16, 3))
+    nx,ny,nz = map(int, np.fromfile(fid, np.uint16, 3))
     start_frames = 0x4c
     fid.seek(start_frames,0)
     frames = np.fromfile(fid, np.uint8, nx*ny*nz).reshape(nz,nx,ny)
